requirement.txt
Django==2.2.6
django-background-tasks==1.2.0
django-cleanup==4.0.0
django-compat==1.0.15
django-cors-headers==3.1.1
django-extra-fields==1.2.4
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
Pillow==6.2.0
pip-autoremove==0.9.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
